Move the resources api location to match our other apis, make it easier to find.
Additional work may be necessary.
