+++ api/resources/helpers.go
@@ -1,7 +1,7 @@
 // Copyright 2015 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package api
+package resources
 
 import (
 	charmresource "github.com/juju/charm/v8/resource"
