+++ api/resources/http.go
@@ -1,7 +1,7 @@
 // Copyright 2016 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package api
+package resources
 
 // TODO(ericsnow) Eliminate the apiserver dependencies, if possible.
