+++ api/resources/private/client/client_test.go
@@ -14,10 +14,10 @@ import (
 	"github.com/juju/testing/filetesting"
 	gc "gopkg.in/check.v1"
 
+	api "github.com/juju/juju/api/resources"
+	"github.com/juju/juju/api/resources/private/client"
 	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/resource"
-	"github.com/juju/juju/resource/api"
-	"github.com/juju/juju/resource/api/private/client"
 	"github.com/juju/juju/resource/resourcetesting"
 )
