+++ apiserver/resources_test.go
@@ -22,12 +22,12 @@ import (
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 
+	api "github.com/juju/juju/api/resources"
 	"github.com/juju/juju/apiserver"
 	apiservererrors "github.com/juju/juju/apiserver/errors"
 	"github.com/juju/juju/apiserver/params"
 	apiservertesting "github.com/juju/juju/apiserver/testing"
 	"github.com/juju/juju/resource"
-	"github.com/juju/juju/resource/api"
 	"github.com/juju/juju/resource/resourcetesting"
 	"github.com/juju/juju/state"
 )
