+++ apiserver/resources_unit.go
@@ -11,9 +11,9 @@ import (
 	"github.com/juju/errors"
 	"github.com/juju/names/v4"
 
+	api "github.com/juju/juju/api/resources"
 	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/resource"
-	"github.com/juju/juju/resource/api"
 	"github.com/juju/juju/state"
 )
