+++ cmd/juju/application/deployer/charm.go
@@ -21,6 +21,7 @@ import (
 	"github.com/juju/juju/api/application"
 	applicationapi "github.com/juju/juju/api/application"
 	commoncharm "github.com/juju/juju/api/common/charm"
+	"github.com/juju/juju/api/resources/client"
 	app "github.com/juju/juju/apiserver/facades/client/application"
 	"github.com/juju/juju/cmd/juju/application/store"
 	"github.com/juju/juju/cmd/juju/application/utils"
@@ -28,7 +29,6 @@ import (
 	"github.com/juju/juju/core/constraints"
 	"github.com/juju/juju/core/devices"
 	"github.com/juju/juju/core/instance"
-	"github.com/juju/juju/resource/api/client"
 	"github.com/juju/juju/resource/resourceadapters"
 	"github.com/juju/juju/storage"
 )
