+++ resource/resourceadapters/deploy.go
@@ -11,9 +11,9 @@ import (
 	"gopkg.in/macaroon.v2"
 
 	"github.com/juju/juju/api/base"
+	"github.com/juju/juju/api/resources/client"
 	resourcecmd "github.com/juju/juju/cmd/juju/resource"
 	"github.com/juju/juju/cmd/modelcmd"
-	"github.com/juju/juju/resource/api/client"
 )
 
 // DeployResourcesFunc is the function type of DeployResources.
